chore: Update webview_flutter and url_launcher dependencies
This commit updates the webview_flutter dependency to the latest version and removes the url_launcher dependency from the plugin list in the generated_plugins.cmake files for both Linux and Windows. Additionally, the pubspec.yaml file is updated to use version 4.0.4 of webview_flutter. These changes ensure that the project is using the most up-to-date versions of these packages and removes any unused dependencies.